Added "say" command + see desc
I added a Client object to represent to player character. This may change in the future to represent client state in relation to the server, and the player will be a character object. Will have to wait and see how that all pans out once we have connections set up.